feat: added recipe click events and observance
