+++ Assets/Scripts/Manager/RecipeShop.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 07d50d1997ec6be4680f78302f5b6b64
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
