+++ Assets/Scripts/Manager/RecipeShop/RecipeClickHandler.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 1b5949b7f8e3c0b498a7c632b035de6f
